ToolStore vs ShoppingCart component separation
